Fix complicated build
