+++ src/x/sync/cpu_amd64.s
