--- src/x/sync/cpu_linux_amd64.go
@@ -1,4 +1,0 @@
-package sync
-
-// CPU is the currently running CPU.
-func CPU() uint64
